Move input/output related code to io package
