+++ mcmd/commands/add.py
@@ -3,12 +3,12 @@ from os import path as os_path
 from pathlib import Path
 
 import mcmd.config.config as config
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post, get, post_files
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.utils.errors import McmdError
 from mcmd.utils.file_helpers import get_file_name_from_path, scan_folders_for_files, select_path
