+++ mcmd/commands/config.py
@@ -1,8 +1,8 @@
 import mcmd.config.config as config
-from mcmd import io
+from mcmd.io import io
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.utils.errors import McmdError
