+++ mcmd/commands/disable.py
@@ -1,9 +1,9 @@
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.molgenis.resources import ensure_resource_exists, ResourceType
