+++ mcmd/commands/enable.py
@@ -1,9 +1,9 @@
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.utils.errors import McmdError
 from mcmd.molgenis.resources import one_resource_exists, ensure_resource_exists, ResourceType
