+++ mcmd/commands/give.py
@@ -5,12 +5,12 @@ principal doesn't exist, the program will terminate.
 """
 from urllib.parse import urljoin
 
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post_form
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.utils.errors import McmdError
 from mcmd.molgenis.principals import ensure_principal_exists, detect_principal_type, PrincipalType
 # =========
