+++ mcmd/commands/set.py
@@ -3,12 +3,12 @@ Alter settings
 """
 import json
 
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import get, put
 from mcmd.command import command
 from mcmd.commands._registry import arguments
-from mcmd.io import highlight
+from mcmd.io.io import highlight
 from mcmd.utils.errors import McmdError
