+++ mcmd/molgenis/auth.py
@@ -9,7 +9,7 @@ from urllib.parse import urljoin
 import requests
 from requests import HTTPError
 
-from mcmd import io
+from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.config import config
 from mcmd.utils.errors import McmdError, MolgenisOfflineError
